add foreign key to notes owner column

+++ src/api/notes/handler.js
@@ -1,6 +1,5 @@
 /* eslint-disable no-unused-vars */
 /* eslint no-underscore-dangle: 0 */
-
 class NotesHandler {
   constructor(service, validator) {
     this._service = service;
@@ -16,8 +15,11 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialId,
+      });
 
     const response = h.response({
       status: 'success',
@@ -30,8 +32,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -40,8 +43,11 @@ class NotesHandler {
     };
   }
 
-  async getNoteByIdHandler(request) {
+ async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -51,11 +57,12 @@ class NotesHandler {
     };
   }
 
-  async putNoteByIdHandler(request) {
+  async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
-    const { id } = request.params;
+    const { id } = request.params; const { id: credentialId } = request.auth.credentials;
 
-    await this._service.editNoteById(id, request.payload);
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.editNoteById(id, request.payload);
 
     return {
       status: 'success',
@@ -63,9 +70,12 @@ class NotesHandler {
     };
   }
 
-  async deleteNoteByIdHandler(request) {
+  async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
-    await this._service.deleteNoteById(id);
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.deleteNoteById(id);
 
     return {
       status: 'success',
